Move files so that includes come from xyz/value_types/
