+++ .github/workflows/cmake.yml
@@ -184,3 +184,14 @@ jobs:
         # Execute tests defined by the CMake configuration.
         # See https://cmake.org/cmake/help/latest/manual/ctest.1.html for more detail
         run: ctest --preset ${{ matrix.configuration }}
+
+      - name: Test FetchContent
+        run: |
+          cmake -S test_package/FetchContent -B test_package/build_fc
+          cmake --build test_package/build_fc --config ${{ matrix.configuration }}
+
+      - name: Test find_package
+        run: |
+          cmake --install build --config ${{ matrix.configuration }} --prefix test_package/install_dir
+          cmake -DCMAKE_PREFIX_PATH=$PWD/test_package/install_dir -S test_package/find_package -B test_package/build_fp
+          cmake --build test_package/build_fp --config ${{ matrix.configuration }}
