+++ benchmarks/polymorphic_benchmark.cc
@@ -38,9 +38,9 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include "experimental/polymorphic_sbo.h"
 #endif  // XYZ_POLYMORPHIC_USES_EXPERIMENTAL_SMALL_BUFFER_OPTIMIZATION
 
-#ifndef XYZ_POLYMORPHIC_H_
-#include "polymorphic.h"
-#endif  // XYZ_POLYMORPHIC_H_
+#ifndef XYZ_VALUE_TYPES_POLYMORPHIC_H
+#include "xyz/value_types/polymorphic.h"
+#endif  // XYZ_VALUE_TYPES_POLYMORPHIC_H
 
 namespace {
