+++ compile_checks/indirect_as_member.cc
@@ -1,7 +1,7 @@
 #include <concepts>
 #include <type_traits>
 
-#include "indirect.h"
+#include "xyz/value_types/indirect.h"
 
 namespace xyz::testing {
