+++ compile_checks/indirect_consteval.cc
@@ -21,7 +21,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // A set of consteval functions to check that constexpr functions can be
 // evaluated at compile time.
 
-#include "indirect.h"
+#include "xyz/value_types/indirect.h"
 
 namespace xyz::testing {
 struct ConstexprHashable {};
