+++ compile_checks/indirect_in_variant.cc
@@ -2,7 +2,7 @@
 #include <type_traits>
 #include <variant>
 
-#include "indirect.h"
+#include "xyz/value_types/indirect.h"
 
 namespace xyz::testing {
