+++ compile_checks/polymorphic_as_member.cc
@@ -1,7 +1,7 @@
 #include <concepts>
 #include <type_traits>
 
-#include "polymorphic.h"
+#include "xyz/value_types/polymorphic.h"
 
 namespace xyz::testing {
