+++ compile_checks/polymorphic_in_variant.cc
@@ -2,7 +2,7 @@
 #include <type_traits>
 #include <variant>
 
-#include "polymorphic.h"
+#include "xyz/value_types/polymorphic.h"
 
 namespace xyz::testing {
