+++ xyz/value_types/indirect.cc
@@ -1,2 +1,2 @@
 // A cc file for indirect to ensure that the header file can be compiled.
-#include "indirect.h"
+#include "xyz/value_types/indirect.h"
