+++ xyz/value_types/indirect.h
@@ -18,8 +18,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ==============================================================================*/
 
-#ifndef XYZ_INDIRECT_H
-#define XYZ_INDIRECT_H
+#ifndef XYZ_VALUE_TYPES_INDIRECT_H
+#define XYZ_VALUE_TYPES_INDIRECT_H
 
 #include <cassert>
 #include <compare>
@@ -485,4 +485,4 @@ struct std::hash<xyz::indirect<T, Alloc>> {
   }
 };
 
-#endif  // XYZ_INDIRECT_H
+#endif  // XYZ_VALUE_TYPES_INDIRECT_H
