+++ xyz/value_types/indirect_cxx14.h
@@ -18,8 +18,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ==============================================================================*/
 
-#ifndef XYZ_INDIRECT_H
-#define XYZ_INDIRECT_H
+#ifndef XYZ_VALUE_TYPES_INDIRECT_H
+#define XYZ_VALUE_TYPES_INDIRECT_H
 
 #include <cassert>
 #include <initializer_list>
@@ -27,7 +27,84 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <type_traits>
 #include <utility>
 
-#include "feature_check.h"
+#ifndef XYZ_VALUE_TYPES_FEATURE_CHECK_H
+#define XYZ_VALUE_TYPES_FEATURE_CHECK_H
+
+// The purpose of the following section is to provide macros checking whether
+// certain C++ features (like std::optional) are available.
+
+//
+// XYZ_HAS_STD_OPTIONAL
+// The macro is defined, when std::optional is available.
+//
+#ifdef XYZ_HAS_STD_OPTIONAL
+#error "XYZ_HAS_STD_OPTIONAL is already defined"
+#endif  // XYZ_HAS_STD_OPTIONAL
+
+#if (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
+#define XYZ_HAS_STD_OPTIONAL
+#endif  //(__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
+        // 201703L)
+
+//
+// XYZ_HAS_STD_IN_PLACE_TYPE_T
+// The macro is defined, when std::in_place_type_t is available.
+//
+
+#ifdef XYZ_HAS_STD_IN_PLACE_TYPE_T
+#error "XYZ_HAS_STD_IN_PLACE_TYPE_T is already defined"
+#endif  // XYZ_HAS_STD_IN_PLACE_TYPE_T
+
+#if (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
+#define XYZ_HAS_STD_IN_PLACE_TYPE_T
+#endif  //(__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
+        // 201703L)
+
+#ifdef XYZ_HAS_STD_IN_PLACE_T
+#error "XYZ_HAS_STD_IN_PLACE_T is already defined"
+#endif  // XYZ_HAS_STD_IN_PLACE_T
+
+#if (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
+#define XYZ_HAS_STD_IN_PLACE_T
+#endif  //(__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
+        // 201703L)
+
+#if (__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 201703L)
+#define XYZ_HAS_TEMPLATE_ARGUMENT_DEDUCTION
+#endif  //(__cplusplus >= 201703L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
+        // 201703L)
+
+//
+// XYZ_HAS_STD_MEMORY_RESOURCE
+// The macro is defined, when the header <memory_resource> and its content is
+// available.
+//
+
+#ifdef __has_include
+#if (__cplusplus >= 201703L) && __has_include(<memory_resource>)
+#include <memory_resource>
+#if __cpp_lib_memory_resource >= 201603L
+#define XYZ_HAS_STD_MEMORY_RESOURCE
+#endif  // __cpp_lib_memory_resource >= 201603L
+#endif  //(__cplusplus >= 201703L) && __has_include(<memory_resource>)
+#endif  //__has_include
+
+//
+// XYZ_HAS_STD_THREE_WAY_COMPARISON
+// The macro is defined when the three-way comparison operator (<=>) is
+// available.
+//
+
+#ifdef XYZ_HAS_STD_THREE_WAY_COMPARISON
+#error "XYZ_HAS_STD_THREE_WAY_COMPARISON is already defined"
+#endif  // XYZ_HAS_STD_THREE_WAY_COMPARISON
+
+#if (__cplusplus >= 202002L) || (defined(_MSVC_LANG) && _MSVC_LANG >= 202002L)
+#define XYZ_HAS_STD_THREE_WAY_COMPARISON
+#endif  //(__cplusplus >= 202002L) || (defined(_MSVC_LANG) && _MSVC_LANG >=
+        // 202002L)
+
+#endif  // XYZ_VALUE_TYPES_FEATURE_CHECK_H
 
 namespace xyz {
 #ifndef XYZ_IN_PLACE_DEFINED
@@ -609,4 +686,4 @@ struct std::hash<xyz::indirect<T, Alloc>> {
   }
 };
 
-#endif  // XYZ_INDIRECT_H
+#endif  // XYZ_VALUE_TYPES_INDIRECT_H
