+++ xyz/value_types/polymorphic.cc
@@ -1,2 +1,2 @@
 // A cc file for polymorphic to ensure that the header file can be compiled.
-#include "polymorphic.h"
+#include "xyz/value_types/polymorphic.h"
