+++ xyz/value_types/polymorphic_cxx14.h
@@ -18,8 +18,8 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ==============================================================================*/
 
-#ifndef XYZ_POLYMORPHIC_H_
-#define XYZ_POLYMORPHIC_H_
+#ifndef XYZ_VALUE_TYPES_POLYMORPHIC_H
+#define XYZ_VALUE_TYPES_POLYMORPHIC_H
 
 #include <cassert>
 #include <initializer_list>
@@ -473,4 +473,4 @@ class polymorphic : private detail::empty_base_optimization<A> {
 
 }  // namespace xyz
 
-#endif  // XYZ_POLYMORPHIC_H_
+#endif  // XYZ_VALUE_TYPES_POLYMORPHIC_H
