+++ xyz/value_types/polymorphic_no_vtable.cc
@@ -1,3 +1,3 @@
 // A cc file for polymorphic_no_vtable to ensure that the header file can be
 // compiled.
-#include "polymorphic_no_vtable.h"
+#include "xyz/value_types/polymorphic_no_vtable.h"
