+++ xyz_value_types-config.cmake.in
@@ -6,3 +6,7 @@ if(NOT TARGET xyz_value_types::xyz_value_types)
     list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}")
     include(${CMAKE_CURRENT_LIST_DIR}/xyz_value_types-export-set.cmake)
 endif()
+
+if(NOT TARGET xyz_value_types::value_types AND TARGET xyz_value_types::xyz_value_types)
+    add_library(xyz_value_types::value_types ALIAS xyz_value_types::xyz_value_types)
+endif()
